Moved some code, added reasons for losing, added word history list
Moved the variables that hold stored attributes of the game into the loop so that they can be reassigned to 'empty' at the start of each game.

Added a 'reason' variable to allow players to know why they lost.

Added a word list to add each player's input word.

Added a function to scan the word list to check if any repeated words were used.